Style form and decouple from dialog
