--- openlibrary/templates/check_ins/dialog.html
@@ -1,71 +1,0 @@
-$def with(id, work_key, edition_key=None)
-
-$# id : str : Unique identifier for the dialog element
-
-$code:
-  year = current_year()
-  months = [
-    _('January'),
-    _('February'),
-    _('March'),
-    _('April'),
-    _('May'),
-    _('June'),
-    _('July'),
-    _('August'),
-    _('September'),
-    _('October'),
-    _('November'),
-    _('December'),
-  ]
-
-<style>
-  dialog {
-    border: none;
-  }
-  dialog::backdrop {
-    opacity: .5;
-    background-color: hsl(0, 0%, 0%);
-  }
-  .check-in-form {
-    display: flex;
-    flex-direction: column;
-  }
-  .check-in-form > div {
-    margin-bottom: 10px
-  }
-  .actions {
-    display: flex;
-    justify-content: space-between;
-  }
-
-</style>
-
-<dialog id="$(id)" class="native-dialog">
-  <form method="dialog" class="olform check-in-form" action="$(work_key)/check-ins" method="post">
-    $if edition_key:
-      <input type="hidden" id="edition-key" name="edition_key" value="$(edition_key)">
-    <div>
-      <label class="label">$_('Start Date:')</label>
-      <select name="year">
-        <option value="">$_('Year')</option>
-        $for i in range(120):
-          <option value="$(year - i)">$(year - i)</option>
-      </select>
-      <select name="month" disabled>
-        <option value="">$_('Month')</option>
-        $for idx, month in enumerate(months):
-          <option value="$(idx + 1)">$month</option>
-      </select>
-      <select name="day" disabled>
-        <option value="">$_('Day')</option>
-        $for i in range(31):
-          <option value="$(i + 1)">$(i + 1)</option>
-      </select>
-    </div>
-    <span class="actions">
-      <button class="close-dialog-btn">$_('Cancel')</button>
-      <button type="submit" class="submit-dialog-btn" disabled>$_('Submit')</button>
-    </span>
-  </form>
-</dialog>
