+++ static/css/components/buttonCta.less
@@ -68,6 +68,11 @@ a.cta-btn {
     border: 2px solid @primary-blue;
     color: @primary-blue;
   }
+  &--shell:disabled {
+    border: 2px solid @light-grey;
+    color: @light-grey;
+    cursor: default;
+  }
   &--sponsor {
     border: 2px solid @link-blue;
     background: lighten(desaturate(@link-blue, 56%), 67%);
@@ -77,6 +82,13 @@ a.cta-btn {
     border: 2px solid @light-grey;
     background: @white;
     color: @black;
+    border-radius: 6px;
+  }
+  &--cancel {
+    padding: 5px;
+    border: 2px solid @red;
+    background: @white;
+    color: @red;
     border-radius: 6px;
   }
