+++ static/css/components/native-dialog.less
@@ -1,0 +1,14 @@
+@import (reference) "../less/colors.less";
+
+.native-dialog {
+  padding: 10px;
+  border: 1px solid @dark-beige;
+  border-radius: 12px;
+  -webkit-box-shadow: 1px 3px 5px @darker-grey;
+  box-shadow: 1px 3px 5px @darker-grey;
+}
+
+.native-dialog::backdrop {
+  opacity: .5;
+  background-color: @black;
+}
