+++ static/css/components/work.less
@@ -388,3 +388,4 @@ div.editionTools {
 }
 
 @import (less) "components/compact-title.less";
+@import (less) "components/check-in.less";
