+++ static/css/js-all.less
@@ -8,6 +8,7 @@
 
 @import (less) "components/header-bar--js.less";
 @import (less) "components/cbox.less";
+@import (less) "components/native-dialog.less";
 @import (less) "components/flash-messages.less";
 @import (less) "components/jquery.autocomplete.less";
